pyproject.toml
[project]
name = "orderservice"
version = "0.1.0"
description = ""
authors = [
    {name = "Dilyaver",email = "yxs.crevier@gmail.com"}
]

requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "aiokafka (>=0.11.0,<0.12.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "bcrypt (==4.0.1)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "alembic (>=1.18.3,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.1.1,<8.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[tool.poetry]
packages = [{include = "orderservice", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: интеграционные тесты, требуют запущенного Kafka (docker-compose.test.yml)",
]

 